Fix use-after-free in (Stateful)PartitionedCallOp.
Before the introduction of StatefulPartitionedCallOp, it was guaranteed that the function would only be partitioned once. That this invariant is no longer true was not reflected in the code prior to this change --- `overlay_lib_` was reset every time the function was re-partitioned. The FunctionLibraryRuntime keeps pointers to overlay libraries, so we need to keep the objects around.

PiperOrigin-RevId: 207902409