Relax out edge testing for now.
Allow a node to flow out to different CondStates. It is not necessarily an error for a node to feed two different CondStates, so relax the testing to simply log a warning for now. The error state is rather due to input to a node given the flow into a merge is the main consideration.

PiperOrigin-RevId: 209500843